simulate sending more passwords than cores
We simulated sending 3 passwords in a 2-core setup, which means that the
core controller needs to wait until a core is ready to receive the 3rd
password.

Note, that we still need to handle and simulate a password overflow
where the core controller cannot keep up with the communication module.